# backendStub/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API reference of The Spoon project.\n \n VERSIONING: The\
    \ version number works in this way: it's composed of two numbers separated by\
    \ a point. The first number identifies the Sprint, so for example the version\
    \ 2.1 of the API reference is related to the Sprint number 2. The second number\
    \ starts from 0 and has to be increased by one each time the file is modified.\
    \ The person who modifies this document is responsible for increasing the number.\n\
    \ \n GROUPING OF ENDPOINTS: The endpoints are grouped by dividing them into customers'\
    \ operations and owners' operations. All the endpoints that are not strictly related\
    \ to customers nor owners remain inside the default group.\n \n AUTHENTICATION/AUTHORIZATION:\
    \ The json web token returned when the login is performed is supposed to be put\
    \ into x-auth-token header in the endpoints that require authorization"
  version: "6.4"
  title: "The Spoon API reference"
...
paths:
  /api/user/customer/register:
    post:
      tags:
      - "customer"
      summary: "Creates customer"
      description: "Creates a new customer profile. This endpoint is used only for\
        \ customer registration.\n The endpoint, if the registration succeeds, returns\
        \ the username of the account as a confirmation."
      operationId: "createCustomer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Customer that needs to register"
        required: true
        schema:
          $ref: "#/definitions/Customer"
      responses:
        "201":
          description: "Customer registered"
          schema:
            $ref: "#/definitions/UsernameAndToken"
        "400":
          description: "Username or email already taken"
      x-swagger-router-controller: "Customer"
...
  /api/user/customer/menu/searchByMenuItem:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Search by menu item"
      description: "Returns all the menus with given menu item (dish/drink) with menuID.\
        \ It also returns the names of the associated restaurants with restaurantID.\n\
        \ The menu items inside every menu are not returned. It will be needed to\
        \ access the endpoint /api/user/customer/menu/{menuID} to get the menu items\
        \ of a specific menu (the menuID passed can be obtained from the response\
        \ of this endpoint, since it returns the menuID of every menu)."
      operationId: "searchByMenuItem"
      produces:
      - "application/json"
      parameters:
      - name: "menuItemName"
        in: "query"
        description: "Name of the desired menu item"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MenuAndRestaurant"
        "404":
          description: "No menu item found"
      x-swagger-router-controller: "Customer"
  /api/user/customer/menu/{menuID}:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Return data of a specific menu"
      description: "Returns all the data about the menu with given menuID, even the\
        \ menu items inside it. The photos of the menu items are saved in the Amazon\
        \ s3 storage, so the links to the cloud storage are also returned. The frontend\
        \ will directly download them from the cloud storage, they won't be sent by\
        \ the backend with this endpoint."
      operationId: "getMenuCustomer"
      produces:
      - "application/json"
      parameters:
      - name: "menuID"
        in: "path"
        description: "ID of the menu to be returned"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/MenuRetrievedToCustomer"
      x-swagger-router-controller: "Customer"
...
  /api/user/customer/review/restaurant:
    get:
      tags:
      - "customer"
      - "Emilio"
      summary: "Return all the restaurants"
      description: "Return all the names of the restaurants and their restaurantIDs."
      operationId: "getRestaurants"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RestaurantIDAndName"
        "401":
          description: "Access denied"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review/restaurant/{restaurantID}/menu:
    get:
      tags:
      - "customer"
      - "Emilio"
      summary: "Return all the menus of given restaurant"
      description: "Return all the names of the menus of the restaurant and their\
        \ menuIDs."
      operationId: "getMenusOfRestaurant"
      produces:
      - "application/json"
      parameters:
      - name: "restaurantID"
        in: "path"
        description: "ID of the restaurant"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MenuIDAndName"
        "401":
          description: "Access denied"
        "404":
          description: "Restaurant with given restaurantID doesn't exist"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review/restaurant/menu/{menuID}/menuItem:
    get:
      tags:
      - "customer"
      - "Emilio"
      summary: "Return all the menu items of given menu"
      description: "Return all the names of the restaurants and their restaurantIDs."
      operationId: "getItemsOfMenu"
      produces:
      - "application/json"
      parameters:
      - name: "menuID"
        in: "path"
        description: "ID of the menu"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MenuItemIDAndName"
        "401":
          description: "Access denied"
        "404":
          description: "Menu with given menuID doesn't exist"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review/restaurant/menu/{menuID}:
    post:
      tags:
      - "customer"
      - "Frikk"
      summary: "Submit a review of the menu"
      description: "Submit a review of the menu with given menuID. The receiptImageID\
        \ is obtained by the frontend when the photo of the receipt is uploaded through\
        \ the dedicated endpoint."
      operationId: "submitReview"
      consumes:
      - "application/json"
      parameters:
      - name: "menuID"
        in: "path"
        description: "ID of the menu"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Review"
        required: true
        schema:
          $ref: "#/definitions/MenuReview"
      responses:
        "201":
          description: "Successful operation"
        "401":
          description: "Access denied"
        "404":
          description: "Menu with given menuID doesn't exist"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Return all the reviews of the customer"
      description: "Return all the reviews of the customer, with their reviewIDs and\
        \ their status (accepted/refused/pending). Returns an empty array if no review\
        \ is found."
      operationId: "getOwnReviews"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MenuReviewWithIDAndStatus"
        "401":
          description: "Access denied"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review/{reviewID}:
    delete:
      tags:
      - "customer"
      - "Marin"
      summary: "Delete a review"
      description: "Delete the review with the given reviewID. This will also delete\
        \ the reviews of the menu items associated to the review with given reviewID."
      operationId: "apiUserCustomerReviewReviewIDDELETE"
      parameters:
      - name: "reviewID"
        in: "path"
        description: "ID of the review"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Access denied"
        "404":
          description: "Review associated to the customer with given reviewID not\
            \ found"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer:
    get:
      tags:
      - "customer"
      - "Marin"
      summary: "Return profile data of the customer"
      description: "Return own data of the logged in customer. This endpoints should\
        \ be used when the frontend has to visualize the profile of the customer."
      operationId: "apiUserCustomerGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/CustomerUsernameAndEmail"
        "401":
          description: "Access denied"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
...
  /api/user/customer/menu/{menuID}/menuItem/{menuItemID}/review:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Return all the reviews of the menu item"
      description: "Return all the reviews of the menu item with given menuItemID\
        \ contained in the menu with given menuID."
      operationId: "apiUserCustomerMenuMenuIDMenuItemMenuItemIDReviewGET"
      produces:
      - "application/json"
      parameters:
      - name: "menuID"
        in: "path"
        description: "ID of the menu"
        required: true
        type: "integer"
      - name: "menuItemID"
        in: "path"
        description: "ID of the menu item"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ReviewsOfMenuItemWithoutOverall"
        "404":
          description: "Menu or menu item not found"
      x-swagger-router-controller: "Customer"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "x-auth-token"
    in: "header"
definitions:
...
  Customer:
    type: "object"
    properties:
      username:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    example:
      username: "xXEmilioXx"
      email: "user@gmail.com"
      password: "123456"
...
  UsernameAndToken:
    type: "object"
    properties:
      username:
        type: "string"
      token:
        type: "string"
    example:
      username: "xXEmilioXx"
      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOjI5LCJpYXQiOjE1NjE5OTg2NjB9.SWYMJXTTM8pe6NQw1QwS-d8Btt6Isuzzk5JtH775uV0"
...
  MenuItemWithID:
    type: "object"
    properties:
      menuItemID:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
      type:
        type: "string"
      priceEuros:
        type: "number"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
      imageLink:
        type: "string"
      rating:
        type: "number"
    example:
      menuItemID: 8
      name: "Spaghetti alla carbonara"
      description: "Fantastic italian dish made of spaghetti, pig cheek, eggs, black\
        \ pepper, pecorino romano"
      type: "dish"
      priceEuros: 10
      tags:
      - name: "Mediterranean"
        color: "#FFBC8C"
      - name: "Pasta"
        color: "#99C99B"
      - name: "Italian"
        color: "#FFBC8C"
      imageLink: "www.cloudStorage.com/Carbonara"
      rating: 4
...
  MenuWithRatingAndID:
    type: "object"
    properties:
      menuID:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
      averagePrice:
        type: "number"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
      rating:
        type: "number"
    example:
      menuID: 2
      name: "Emilio's menu of the day"
      description: "Our special menu of today"
      averagePrice: 9.5
      tags:
      - name: "Italian"
        color: "#FFBC8C"
      - name: "Mediterranean"
        color: "#FFBC8C"
      rating: 5
  MenuAndRestaurant:
    type: "object"
    properties:
      restaurantData:
        $ref: "#/definitions/MenuAndRestaurant_restaurantData"
      menu:
        $ref: "#/definitions/MenuWithRatingAndID"
    example:
      restaurantData:
        restaurantName: "Emilio's Pizza"
        restaurantImageLink: "www.cloudStorage.com/Restaurant"
        distance: 10
      menu:
        menuID: 2
        name: "Emilio's menu of the day"
        description: "Our special menu of today"
        averagePrice: 9.5
        tags:
        - name: "Italian"
          color: "#FFBC8C"
        - name: "Mediterranean"
          color: "#FFBC8C"
        rating: 5
...
  MenuRetrievedToCustomer:
    type: "object"
    properties:
      restaurant:
        $ref: "#/definitions/MenuRetrievedToCustomer_restaurant"
      menuName:
        type: "string"
      description:
        type: "string"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
      menuRating:
        type: "number"
      menuItems:
        type: "array"
        items:
          $ref: "#/definitions/MenuItemWithID"
    example:
      restaurant:
        restaurantName: "Emilio's Pizza"
        address: "Piazzale Susa"
        city: "Milan"
        country: "Italy"
        latitude: 45.46883460000000098943928605876863002777099609375
        longitude: 9.223411399999999815690898685716092586517333984375
      menuName: "Emilio's menu of the day"
      description: "Our special menu of today"
      tags:
      - name: "Italian"
        color: "#FFBC8C"
      - name: "Mediterranean"
        color: "#FFBC8C"
      menuRating: 5
      menuItems:
      - menuItemID: 20
        name: "Spaghetti alla carbonara"
        description: "Fantastic italian dish made of spaghetti, pig cheek, eggs, black\
          \ pepper, pecorino romano"
        type: "dish"
        priceEuros: 10
        tags:
        - name: "Mediterranean"
          color: "#FFBC8C"
        - name: "Pasta"
          color: "#99C99B"
        - name: "Italian"
          color: "#FFBC8C"
        - name: "Dinner"
          color: "#99C99B"
        imageLink: "www.cloudStorage.com/Carbonara"
        rating: 4.5
      - menuItemID: 21
        name: "Polpette al sugo"
        description: "Meatballs with tomato sauce"
        type: "dish"
        priceEuros: 7
        tags:
        - name: "Mediterranean"
          color: "#FFBC8C"
        - name: "Meat"
          color: "#FFBC8C"
        - name: "Italian"
          color: "#FFBC8C"
        imageLink: "www.cloudStorage.com/Meatballs"
        rating: 4
  Tag:
    type: "object"
    properties:
      name:
        type: "string"
      color:
        type: "string"
    example:
      name: "Mediterranean"
      color: "#FFBC8C"
  RestaurantIDAndName:
    type: "object"
    properties:
      name:
        type: "string"
      restaurantID:
        type: "integer"
    example:
      name: "Emilio's Pizza"
      restaurantID: 520
  MenuIDAndName:
    type: "object"
    properties:
      name:
        type: "string"
      menuID:
        type: "integer"
    example:
      name: "Emilio's menu of the day"
      menuID: 420
  MenuItemIDAndName:
    type: "object"
    properties:
      name:
        type: "string"
      menuItemID:
        type: "integer"
    example:
      name: "Pasta alla carbonara"
      menuItemID: 802
  MenuReview:
    type: "object"
    properties:
      serviceRating:
        type: "number"
      qualityOverPriceRating:
        type: "number"
      date:
        type: "string"
      receiptImageID:
        type: "integer"
      menuItemsReviews:
        type: "array"
        items:
          $ref: "#/definitions/MenuItemReview"
    example:
      serviceRating: 4.5999999999999996447286321199499070644378662109375
      qualityOverPriceRating: 4.70000000000000017763568394002504646778106689453125
      date: "2019-12-01"
      receiptImageID: 67
      menuItemsReviews:
      - menuItemID: 34
        rating: 4.0999999999999996447286321199499070644378662109375
        content: "That was delicious!"
      - menuItemID: 58
        rating: 0.5
        content: "It was an insult, I will never eat that trash again in my whole\
          \ life"
  MenuItemReview:
    type: "object"
    properties:
      menuItemID:
        type: "integer"
      rating:
        type: "number"
      content:
        type: "string"
  MenuReviewWithIDAndStatus:
    type: "object"
    properties:
      menuReviewID:
        type: "integer"
      menuID:
        type: "integer"
      menuName:
        type: "string"
      restaurantName:
        type: "string"
      serviceRating:
        type: "number"
      qualityOverPriceRating:
        type: "number"
      date:
        type: "string"
      receiptImageID:
        type: "integer"
      status:
        type: "string"
      menuItemsReviews:
        type: "array"
        items:
          $ref: "#/definitions/MenuItemReview"
    example:
      menuReviewID: 34342
      menuID: 343
      menuName: "Fish menu"
      restaurantName: "Emilio's Restaurant"
      serviceRating: 4.5999999999999996447286321199499070644378662109375
      qualityOverPriceRating: 4.70000000000000017763568394002504646778106689453125
      date: "2019-12-01"
      receiptImageID: 67
      status: "accepted"
      menuItemsReviews:
      - menuItemID: 34
        rating: 4.0999999999999996447286321199499070644378662109375
        content: "That was delicious!"
      - menuItemID: 58
        rating: 0.5
        content: "It was an insult, I will never eat that trash again in my whole\
          \ life"
  CustomerUsernameAndEmail:
    type: "object"
    properties:
      username:
        type: "string"
      email:
        type: "string"
    example:
      username: "emilio_imperiali"
      email: "emilioimperiali@mail.it"
...
  ReviewsOfMenuItemWithoutOverall:
    type: "object"
    properties:
      username:
        type: "string"
      rating:
        type: "number"
      content:
        type: "string"
      date:
        type: "string"
    example:
      username: "Janine"
      rating: 5
      content: "Best pizza I have tasted in ages!"
      date: "2019-12-01"
...
  MenuAndRestaurant_restaurantData:
    type: "object"
    properties:
      restaurantName:
        type: "string"
      restaurantImageLink:
        type: "string"
      distance:
        type: "number"
  MenuRetrievedToCustomer_restaurant:
    type: "object"
    properties:
      restaurantName:
        type: "string"
      address:
        type: "string"
      city:
        type: "string"
      country:
        type: "string"
      latitude:
        type: "number"
      longitude:
        type: "number"
...